+++ timpute/test/test_dopt.py
@@ -1,7 +1,7 @@
 import numpy as np
-from timpute.direct_opt import perform_DO
-from timpute.cmtf import calcR2X
-from timpute.figures import generateTensor
+from ..method_DO import perform_DO
+from ..impute_helper import calcR2X
+from ..attempt_figures import generateTensor
 from timpute.tracker import tracker
 from timpute.decomposition import Decomposition
 from timpute.common import getSetup
